requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.18.22
botocore==1.21.22
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==36.0.0
dj-database-url==0.5.0
Django==3.2.6
django-appconf==1.0.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-select2==7.7.1
django-sendgrid-v5==0.9.1
django-storages==1.11.1
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.2
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.21
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python-http-client==3.3.2
pytz==2021.1
razorpay==1.2.0
reportlab==3.6.1
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
stripe==2.60.0
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
